+++ package.json
@@ -19,6 +19,7 @@
     "jwt-decode": "^3.1.2",
     "public-ip": "^4.0.4",
     "react": "^17.0.2",
+    "react-color": "^2.19.3",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
     "react-icons": "^4.3.1",
