+++ src/Components/Admin/SwitchPage/index.jsx
@@ -1,13 +1,18 @@
+import { useAuth } from "../../../hooks/useAuth"
+
 import { About } from "./Pages/About"
 import { Convention } from "./Pages/Convention"
 import { Cults } from "./Pages/Cults"
 import { Donate } from "./Pages/Donate"
+import { AddChurch } from "./Pages/AddChurch"
 import { MyChurch } from "./Pages/MyChurch"
 
 export function SwitchPage({ page }){
+  const { user } = useAuth();
+  console.log(user.church)
   switch (page){
     case 'Minha Igreja':
-      return <MyChurch/>
+      return user.church != null ? <MyChurch/> : <AddChurch/>
     case 'Sobre':
       return <About/>
     case 'Cultos':
